moved feature columns base class to base.py
