+++ easyflow/feature_encoders/__init__.py
@@ -1,7 +1,7 @@
 """Keras and tensorflow feature column pipeline
 """
 
-from .base import BaseFeatureColumnEncoder
+from .base import BaseFeatureColumnEncoder, BaseCategoricalFeatureColumnEncoder, get_unique_vocabulary
 from .categorical_encoders import CategoricalFeatureEncoder, EmbeddingFeatureEncoder,\
     CategoryCrossingFeatureEncoder
 from .numerical_encoders import NumericalFeatureEncoder, BucketizedFeatureEncoder
@@ -9,6 +9,7 @@ from .pipeline import FeatureColumnTransformer, FeatureUnionTransformer
 
 
 __all__ = ["BaseFeatureColumnEncoder",
+           "BaseCategoricalFeatureColumnEncoder",
            "CategoricalFeatureEncoder",
            "EmbeddingFeatureEncoder",
            "CategoryCrossingFeatureEncoder",
