+++ easyflow/feature_encoders/base.py
@@ -55,3 +55,55 @@ class BaseFeatureColumnEncoder(ABC):
             (dict, list): Keras inputs for each feature and list of encoders
         """
         return self.create_inputs(dataset, features), self.encode(dataset, features)
+
+
+def get_unique_vocabulary(dataset=None, features=None):
+    """Get feature vocabulary list
+
+    Args:
+        dataset (tf.data.Dataset): Features Data to apply encoder on.
+        features (list): list of feature names
+
+    Returns:
+        (dict): dictionary containing list of unique vocabulary values for each feature
+    """
+    if hasattr(dataset, '_batch_size'):
+        # unbatch dataset
+        dataset = dataset.unbatch()
+
+    if isinstance(dataset.element_spec, tuple):
+        map_func = lambda x, y: x[feature]
+    else:
+        map_func = lambda x: x[feature]
+
+    feature_vocab_list = {}
+    for feature in features:
+        feature_ds = dataset.map(map_func)\
+                            .apply(tf.data.experimental.unique())\
+                            .as_numpy_iterator()
+        uniq_vocab = list(feature_ds)
+        if all(map(lambda x: isinstance(x, bytes), uniq_vocab)):
+            # map bytes to ensure objects are serializable when saving model
+            uniq_vocab = [str(value, 'utf-8') for value in uniq_vocab]
+        feature_vocab_list[feature] = tf.feature_column.categorical_column_with_vocabulary_list(feature, uniq_vocab)
+    return feature_vocab_list
+
+
+class BaseCategoricalFeatureColumnEncoder(BaseFeatureColumnEncoder):
+    """Base class for categorical type encoders
+    """
+    def __init__(self, feature_transformer=None, **kwargs):
+        super().__init__(feature_transformer=feature_transformer, **kwargs)
+
+    def encode(self, dataset=None, features=None):
+        """Apply feature encoding that requires unique vocabulary as input
+
+        Args:
+            dataset (tf.data.Dataset): Features Data to apply encoder on.
+            features (list): list of feature names
+
+        Returns:
+            (list): list of encoded features
+        """
+        feature_vocab_list = get_unique_vocabulary(dataset, features)
+        return [self.feature_transformer(feature_vocab_list[feature], **self.kwargs) for feature in features]
